+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_06_05_120000) do
+ActiveRecord::Schema[8.1].define(version: 2026_06_05_120003) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -109,8 +109,8 @@ ActiveRecord::Schema[8.1].define(version: 2026_06_05_120000) do
     t.datetime "created_at", null: false
     t.boolean "default", default: false, null: false
     t.string "handle", null: false
+    t.text "instructions"
     t.string "name", null: false
-    t.string "source_path", null: false
     t.datetime "updated_at", null: false
     t.bigint "workspace_id", null: false
     t.index ["workspace_id", "handle"], name: "index_transformer_profiles_on_workspace_id_and_handle", unique: true